requirements.txt
scikit-learn==1.3.0
catboost==1.2
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.4
imbalanced-learn==0.11.0
skrebate==0.62
Boruta==0.3
shap==0.42.1
simulated-annealing==0.4.0
costcla==0.6
umap-learn==0.5.3
yellowbrick==1.5
numba==0.57.1
jupyterlab==4.0.5
pylint==2.17.5
yapf==0.40.1
scikit-multilearn==0.2.0
# joblib==1.3.2